Better diagnostics for map iteration while mutating (#10)
If one obtains a `MutableMap.Entry` from a mutable map / builder, and
then removes the underlying value from the map, we currently return
`null` from the entry's `value` property, rather than throwing. Either
behavior is allowed by the `MutableMap.Entry` spec, which specifies
`IllegalStateException` *or* undefined behavior in this case - but it's
much easier to debug this if an exception is thrown.