Introduce TreapList (#12)
`TreapList` is a Treap-based `PersistentList` implementation which
significantly outperforms both `ArrayList` and the reference
`PersistentList` implementation from `kotlinx.collections.immutable` for
most operations. Nearly every `PersistentList` operation has a log-time
implementation, including things like inserting one list in the middle
of another list.

`TreapList` is notably a rather efficient persistent Deque, with
O(log(N)) insertion and removal at both ends of the list.

## Design

`TreapList` is organized as a Treap, with left/right being determined by
the order of the elements in the list, and with randomly-assigned node
priorities to balance the tree. This is different from how we assign
priorities in `TreapMap`/`TreapSet`, where we hash the element values to
get the priority; because `TreapList` can contain multiple elements with
the same value, we must randomly assign priorities to ensure that, say,
a list of 1,000,000 entries with the same value won't just end up being
a linked list. If we randomly assigned priorities in the set/map
implementations, it would complicate unions, intersections, merges, etc.
- but those operations do not apply to Lists, so random priorities are
fine here.

Also unlike the set/map implementations, each `TreapList` node tracks
the size of the sublist represented by that node, enabling log-time
indexing into the list.

`TreapList` is a little like a probabilistically balanced
[Rope](https://en.wikipedia.org/wiki/Rope_(data_structure)), but as an
arbitrary list.

## Performance

For most operations, `TreapList` outperforms `ArrayList` (when used as
an immutable list), as well as the reference `PersistentList`
implementation from `kotlinx.collections.immutable`. The most notable
exception is the `get` operation, which for `ArrayList` is a simple
array indexing operation:


![image](https://github.com/Certora/collections/assets/7407587/b6b1d85d-a2f7-44c5-b403-6495e816f87f)

`TreapList` generally matches or beats `ArrayList` when adding a single
element to the end of the list (producing a new list), but is beaten by
`kotlinx.collections.immutable`:


![image](https://github.com/Certora/collections/assets/7407587/8f073ef4-fb96-4472-8b02-f0a73929930a)

However, for most other list operations, `TreapList` wins handily. For
example, insertion of a single item at the front of the list:


![image](https://github.com/Certora/collections/assets/7407587/b28fd2ac-8cfc-466e-9d21-1733746e93dd)

Replacing the value at a given index:


![image](https://github.com/Certora/collections/assets/7407587/6020ae6a-d93b-4516-ac50-273dc3b40208)

Or appending one list to another list:


![image](https://github.com/Certora/collections/assets/7407587/b0d66ed5-09d3-4781-97d4-c579b4ffea09)

More benchmarks are available; you can run `gradlew listBenchmark` to
run them all.

## Memory Usage

One disadvantage of `TreapList` vs the alternatives is that a single
`TreapList` of a given size is quite a lot larger than the equivalent
`ArrayList` or `kotlinx.collections.immutable` list:


![image](https://github.com/Certora/collections/assets/7407587/f8b6f575-b237-4913-b46b-ce29909e0aea)

However, `TreapList` is better able to re-use allocations between
"versions" of a list. For example, consider the scenario where we add
one item at a time, in the middle of the list, retaining all
intermediate results, and adding up the total size:


![image](https://github.com/Certora/collections/assets/7407587/42cd6dd4-5433-44fe-a196-86bce863eb42)

We can see that in this extreme case, `TreapList` uses much less memory
than the other alternatives.

Real heap usage will of course depend on how much the specific use case
is able to take advantage of the increased allocation sharing.