Intersection mode for `merge` (#21)
#19 added the `union` and `intersect` methods to `TreapMap`, which are
more efficient than `merge` in the case where we only care about merging
the entries that share keys. However, we also have a need for a function
that can apply the full `merge` semantics (including removal of keys),
but optimized for the intersection case. This PR adds `merge` and
`parallelMerge` overloads that take a mode argument; the two modes are
`UNION` (the existing behavior) and `INTERSECTION` (only includes common
keys in the merge).

These are separate overloads, rather than simply adding a new optional
parameter to the existing ones, because adding a new parameter in the
middle of the parameter list would be a breaking change.