`TreapMap.keys` as `TreapSet` (#20)
Currently it is quite expensive to do operations like this:

```
set2 = map1.keys intersect set1
```

This is because the `keys` property on `TreapMap` is an `ImmuableSet`,
not a `TreapSet`, so the intersection operation must convert it to a
`TreapSet` first, using the generic `Set`->`TreapSet` conversion, which
is rather expensive.

This is a shame, because a `TreapMap<K, V>` of course is already
structured the same way as a `TreapSet<K>`, so the conversion to
`TreapSet` should be very cheap.

This PR redefines the `keys` property as a `TreapSet`, and defines
efficient implementations of this. Depending on usage, we can sometimes
avoid the conversion to `TreapSet` altogether; if not, we can do the
conversion as a straightforward projection of the existing `TreapMap`
structure, avoiding most of the overhead of the generic conversion.

We also add `single` and `singleOrNull` methods to `TreapMap`, which are
useful in their own right, and can be used to avoid the conversion to
`TreapSet` in some cases.